Parameters and function types update

+++ DiscordOverlayHooking/pattern_scanner.h
@@ -5,6 +5,8 @@
 #include <TlHelp32.h>
 #include <Psapi.h>
 
+// This class manages the pattern scanning part
+
 class PatternScanning
 {
 public:
@@ -12,7 +14,7 @@ public:
 	PatternScanning(HANDLE hProcess, HMODULE hModule);
 
 	uintptr_t PatternScan(std::uint8_t* pattern, char* mask);
-	unsigned FindPattern(std::uint8_t* moduleContent, unsigned int moduleSize, std::uint8_t* pattern, char* mask);
+	uintptr_t FindPattern(std::uint8_t* moduleContent, uintptr_t moduleSize, std::uint8_t* pattern, char* mask);
 
 private:
 	HANDLE _hProcess;
